+++ shared/Cargo.toml
@@ -12,7 +12,6 @@ blake2b_simd = "1.0.0"
 thiserror = "1.0"
 num-traits = "0.2"
 num-derive = "0.3"
-num_enum = "0.5"
 num-bigint = "0.4"
 num-integer = "0.1"
 data-encoding = "2.3.2"
